Added Mersenne-Twister random number generator (RNG) EvtMTRandomEngine, which requires c++11 compiler features, enabled by the configure script when the compiler version >= gcc 4.7.
Added a phase-space test validation/genPHSP.sh and PhaseSpacePlots.C to
visually check the flatness of Dalitz plots in order to ensure that the
RNG is not producing biased results that depend on particle ordering.

Added the models EvtbsToLLLLAmp and EvtbsToLLLLHyperCP for
B0_q -> l+ l- l+ l- decays (SM and one supersymmetric scenario).